refactor(viz): move textAlign outside of TextNode and Style.
