+++ viz/d2v-viz-common/src/test/kotlin/io/data2viz.viz/TestTextNode.kt
@@ -1,0 +1,19 @@
+package io.data2viz.viz
+
+import io.data2viz.test.*
+import kotlin.test.Test
+
+class TestTextNode: TestBase(){
+
+
+    @Test
+    fun accessibility(){
+        val textNode = TextNode().apply {
+            textAlign = textAlign(TextAnchor.END)
+        }
+
+        textNode.style.baseline shouldBe TextAlignmentBaseline.BASELINE
+        textNode.style.anchor shouldBe TextAnchor.END
+
+    }
+}
